refactored gencmd
